fixed types and added automatic type checking

+++ rasa/core/training/interactive.py
@@ -26,7 +26,7 @@ from rasa.core.actions.action import (
     default_action_names,
     UTTER_PREFIX,
 )
-from rasa.core.channels import UserMessage
+from rasa.core.channels.channel import UserMessage
 from rasa.core.channels.channel import button_to_string, element_to_string
 from rasa.core.constants import (
     DEFAULT_SERVER_FORMAT,
@@ -324,7 +324,7 @@ async def _ask_questions(
 
 def _selection_choices_from_intent_prediction(
     predictions: List[Dict[Text, Any]]
-) -> List[Dict[Text, Text]]:
+) -> List[Dict[Text, Any]]:
     """"Given a list of ML predictions create a UI choice list."""
 
     sorted_intents = sorted(predictions, key=lambda k: (-k["confidence"], k["name"]))
@@ -428,7 +428,7 @@ async def _request_intent_from_user(
         intent_name = await _request_free_text_intent(sender_id, endpoint)
         return {"name": intent_name, "confidence": 1.0}
     # returns the selected intent with the original probability value
-    return next((x for x in predictions if x["name"] == intent_name), None)
+    return next((x for x in predictions if x["name"] == intent_name))
 
 
 async def _print_history(sender_id: Text, endpoint: EndpointConfig) -> None:
@@ -526,7 +526,7 @@ def _chat_history_table(evts: List[Dict[Text, Any]]) -> Text:
 
         else:
             e = Event.from_parameters(evt)
-            if e.as_story_string():
+            if e is not None and e.as_story_string():
                 bot_column.append(wrap(e.as_story_string(), bot_width(table)))
 
     if bot_column:
@@ -546,7 +546,7 @@ def _slot_history(tracker_dump: Dict[Text, Any]) -> List[Text]:
     """Create an array of slot representations to be displayed."""
 
     slot_strs = []
-    for k, s in tracker_dump.get("slots").items():
+    for k, s in tracker_dump.get("slots", {}).items():
         colored_value = cliutils.wrap_with_color(
             str(s), color=rasa.cli.utils.bcolors.WARNING
         )
@@ -603,7 +603,7 @@ async def _ask_if_quit(sender_id: Text, endpoint: EndpointConfig) -> bool:
 
 async def _request_action_from_user(
     predictions: List[Dict[Text, Any]], sender_id: Text, endpoint: EndpointConfig
-) -> (Text, bool):
+) -> Tuple[Text, bool]:
     """Ask the user to correct an action prediction."""
 
     await _print_history(sender_id, endpoint)
@@ -716,7 +716,7 @@ def _collect_messages(evts: List[Dict[Text, Any]]) -> List[Message]:
 
     for evt in evts:
         if evt.get("event") == UserUttered.type_name:
-            data = evt.get("parse_data")
+            data = evt.get("parse_data", {})
 
             for entity in data.get("entities", []):
 
@@ -929,6 +929,10 @@ async def _correct_wrong_nlu(
     """A wrong NLU prediction got corrected, update core's tracker."""
 
     latest_message = latest_user_message(evts)
+
+    if latest_message is None:
+        raise Exception("Failed to correct NLU data. User message not found.")
+
     corrected_events = all_events_before_latest_user_msg(evts)
 
     latest_message["parse_data"] = corrected_nlu
@@ -938,7 +942,7 @@ async def _correct_wrong_nlu(
     await send_message(
         endpoint,
         sender_id,
-        latest_message.get("text"),
+        latest_message.get("text", ""),
         latest_message.get("parse_data"),
     )
 
@@ -1076,7 +1080,7 @@ def _as_md_message(parse_data: Dict[Text, Any]) -> Text:
     from rasa.nlu.training_data.formats import MarkdownWriter
 
     if parse_data.get("text", "").startswith(INTENT_MESSAGE_PREFIX):
-        return parse_data.get("text")
+        return parse_data["text"]
 
     if not parse_data.get("entities"):
         parse_data["entities"] = []
@@ -1140,9 +1144,11 @@ async def _validate_nlu(
 
     tracker = await retrieve_tracker(endpoint, sender_id, EventVerbosity.AFTER_RESTART)
 
-    latest_message = latest_user_message(tracker.get("events", []))
+    latest_message = latest_user_message(tracker.get("events", [])) or {}
 
-    if latest_message.get("text").startswith(INTENT_MESSAGE_PREFIX):
+    if latest_message.get("text", "").startswith(  # pytype: disable=attribute-error
+        INTENT_MESSAGE_PREFIX
+    ):
         valid = _validate_user_regex(latest_message, intents)
     else:
         valid = await _validate_user_text(latest_message, endpoint, sender_id)
@@ -1531,7 +1537,7 @@ def run_interactive_learning(
 
     if not skip_visualization:
         p = Process(target=start_visualization, args=("story_graph.dot",))
-        p.deamon = True
+        p.daemon = True
         p.start()
     else:
         p = None
@@ -1554,6 +1560,6 @@ def run_interactive_learning(
 
     _serve_application(app, stories, finetune, skip_visualization)
 
-    if not skip_visualization:
+    if not skip_visualization and p is not None:
         p.terminate()
         p.join()
